Rename SegmentExecution to LegacySegmentExecution
